w1-eeprom: Make use of static keyword in drivers
Both the ds24xxx and ds2502 have a family_supported array which is local
to the driver but was not marked as static. Mark this as static as both
best practice and to allow both to be built in "allyesconfig" for
example.

Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
(cherry picked from commit 7781f1d832bbf9461b82d48301091b15fcf1be14)