luks: test: Check file access in the decryption tests
The current tests check that decryption happens but don't go so far as
reading a file. Add the logic for this. Since this currently fails for
LUKSv2, leave the last part of that test out for now.

Use direct filesystem calls for LUKSv2 since it is easier to check for
the current error.

Add more information on the errors returned by luks_unlock() so we can
check for the correct one.

Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <simon.glass@canonical.com>