tkey: Support loading an app with a user-supplied secret
One useful feature a the TKey is the ability to set up its app with an
key modified by a user-supplied secret. Add support for this.

Take this opportunity to make the API more expo-friendly by allowing
loading of the TKey to take place iteratively. The TKey runs fairly
slowly (about 60Kbaud) and loading an app takes 6 seconds or so.

Signed-off-by: Simon Glass <simon.glass@canonical.com>