bootctl: Enable the tests
Enable the bootctl tests (only for the sandbox board) so we can keep
this functionality working.

Series-to: concept
Cover-letter:
bootctl: Continue development with TKey functionality
This series integrates the TKey disk-unlock features into the bootctl
UI, as a demonstration of how this might work. The user is prompted for
a passphrase, which is then used as a user-supplied secret (USS) for the
TKey.

This series includes support for using a pre-derived master key, so that
the TKey emulator can be used in tests.

Future work will continue this effort.
END

Signed-off-by: Simon Glass <simon.glass@canonical.com>