test: fit: Add pytest for mkimage output
Add a new pytest to check the output of 'mkimage -l' on the test FIT.

Only check the first 80 chars as the signature lines can be very long.

The timestamps are quite annoying, since we need the test to pass on
whatever CI machine is used. Obtain the first timestamp and use that to
check the rest, after some sanity checks.

Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <simon.glass@canonical.com>